+++ Project_2-master/routes/web.php
@@ -101,7 +101,8 @@ Route::get('bib/{id}', [BibtexController::class, 'getbib'])->name('bibtex');
 Route::get('/highlight', [ShowAllNews::class, 'index'])->name('highlight.index');
 
 //สำหรับหน้ารายละเอียดข่าว
-Route::get('/news/{id}',[ReadNewsController::class,'index'])->name('news.index');
+Route::get('/news/{id}', [ReadNewsController::class, 'index'])->name('news.details');
+
 
 
 
@@ -162,6 +163,11 @@ Route::group(['middleware' => ['auth', 'PreventBackHistory']], function () {
     Route::post('/tag/store', [ManageHighlight::class, 'storeTag'])->name('tag.store');
     Route::put('/tag/update', [ManageHighlight::class, 'updateTag'])->name('tag.update');
     Route::delete('/tag/delete', [ManageHighlight::class, 'destroyTag'])->name('tag.delete');
+
+    Route::get('/edit-highlight/{id}', [ManageHighlight::class, 'editHighlight'])->name('highlight.edit');
+    Route::get('/preview-highlight/{id}', [ManageHighlight::class, 'previewHighlight'])->name('highlight.preview');
+    Route::put('/update-highlight/{id}/', [ManageHighlight::class, 'updateHighlight'])->name('highlight.update');
+    Route::post('/news/{id}/publish', [ManageHighlight::class, 'publish'])->name('highlight.publish');
 });